refactor: Separate logger initialization into bootstrap-logger
- Move logger setup from backend.ts, lambda.ts, main.ts, all_exception.filter.ts and db.ts to bootstrap-logger.
- Improve code modularity and maintainability.
