+++ src/main.ts
@@ -1,5 +1,6 @@
+import { logger } from "@/infrastructure/logger/bootstrap-logger";
+
 import { createBackend } from "./backend";
-import { logger } from "./infrastructure/logger";
 
 const bootstrap = async () => {
     try {
